requirements.txt
asgiref
certifi
charset-normalizer
Django
django-admin-autocomplete-filter
django-admin-rangefilter
django-cors-headers
django-filter
djangorestframework
drf-yasg
et-xmlfile
idna
inflection
openpyxl
packaging
pycryptodome
pyTelegramBotAPI
python-dotenv
pytz
PyYAML
requests
sqlparse
uritemplate
urllib3
